+++ libs/organization/src/lib/organization/pages/member/member.component.ts
@@ -3,11 +3,11 @@ import { Observable } from 'rxjs';
 import { OrganizationQuery } from '../../+state/organization.query';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { PermissionsQuery, UserRole, PermissionsService } from '../../../permissions/+state';
-import { UserService } from '@blockframes/user/+state/user.service';
 import { UserQuery } from '@blockframes/user/+state/user.query';
 import { InvitationService } from '@blockframes/invitation/+state/invitation.service';
 import { Invitation } from '@blockframes/invitation/+state/invitation.model';
 import { OrganizationMember } from '@blockframes/user/+state/user.model';
+import { OrganizationService } from '@blockframes/organization/+state';
 
 @Component({
   selector: 'member-edit',
@@ -37,7 +37,7 @@ export class MemberComponent implements OnInit {
     private permissionQuery: PermissionsQuery,
     private permissionService: PermissionsService,
     private userQuery: UserQuery,
-    private userService: UserService
+    private orgService: OrganizationService,
   ) { }
 
   ngOnInit() {
@@ -96,7 +96,7 @@ export class MemberComponent implements OnInit {
 
   public removeMember(uid: string) {
     try {
-      this.userService.removeMember(uid);
+      this.orgService.removeMember(uid);
       this.snackBar.open('Member removed.', 'close', { duration: 2000 });
     } catch (error) {
       this.snackBar.open(error.message, 'close', { duration: 2000 });
